refactor: controller, service 폴더 구조 정리
