+++ src/main/java/com/hufs/algoing/recommendation/service/RecommendService.java
@@ -8,34 +8,57 @@ import com.hufs.algoing.problem.entity.Problem;
 import com.hufs.algoing.problem.entity.SubmittedProblem;
 import com.hufs.algoing.problem.repository.ProblemRepository;
 import com.hufs.algoing.problem.repository.SubmittedProblemRepository;
+import com.hufs.algoing.recommendation.algorithm.DailyRecommendAlgorithm;
+import com.hufs.algoing.recommendation.algorithm.IncProblemRecommendAlgorithm;
 import com.hufs.algoing.recommendation.algorithm.WeaknessRecommendAlgorithm;
+import com.hufs.algoing.recommendation.dto.DailyRecommendDTO;
+import com.hufs.algoing.recommendation.dto.IncProblemRecommendDTO;
 import com.hufs.algoing.recommendation.dto.WeaknessRecommendDTO;
 import com.hufs.algoing.review.entity.Review;
 import com.hufs.algoing.review.repository.ReviewRepository;
 import com.hufs.algoing.user.entity.User;
 import com.hufs.algoing.user.repository.UserRepository;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class WeaknessRecommendService {
+public class RecommendService {
 
-    private final UserRepository userRepository;
     private final ProblemRepository problemRepository;
     private final SubmittedProblemRepository submittedProblemRepository;
+    private final UserRepository userRepository;
     private final ReviewRepository reviewRepository;
     private final AISolvedRepository aisolvedRepository;
+    private final IncProblemRecommendAlgorithm incProblemRecommendAlgorithm;
 
+    public List<DailyRecommendDTO> getDailyRecommendations(Long userId) {
+
+        // 유저 정보 가져오기
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException(ErrorStatus.USER_NOT_FOUND));
+
+        // 유저가 푼 문제 가져오기
+        List<SubmittedProblem> solvedProblems = submittedProblemRepository.findByUserId(user);
+
+        // 전체 문제 가져오기
+        List<Problem> allProblems = problemRepository.findAll();
+
+        // 추천 알고리즘 호출
+        List<DailyRecommendDTO> recommendations = DailyRecommendAlgorithm.recommend(user, allProblems, solvedProblems);
+
+        for (DailyRecommendDTO dto : recommendations) {
+        }
+        return recommendations;
+    }
 
     public List<WeaknessRecommendDTO> getWeaknessRecommendations(Long userId) {
 
         //유저 정보 가져오기
-        User user=userRepository.findById(userId)
-                .orElseThrow(()-> new UserNotFoundException(ErrorStatus.USER_NOT_FOUND));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException(ErrorStatus.USER_NOT_FOUND));
         System.out.println(user);
 
         //전체 문제 가져오기
@@ -48,7 +71,7 @@ public class WeaknessRecommendService {
         List<Review> allReviews = reviewRepository.findAll();
 
         //AI가 푼 문제 가져오기
-        List<AISolved> allAiSolved =aisolvedRepository.findAll();
+        List<AISolved> allAiSolved = aisolvedRepository.findAll();
 
         // 추천 알고리즘 호출
         List<WeaknessRecommendDTO> recommendations = WeaknessRecommendAlgorithm.recommend(user, allReviews, allAiSolved, allSolvedProblems, allProblems);
@@ -56,6 +79,17 @@ public class WeaknessRecommendService {
         }
         return recommendations;
 
+
+    }
+
+    public List<IncProblemRecommendDTO> getIncProblemRecommendation(Long userId) {
+
+        //유저 정보 가져오기
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException(ErrorStatus.USER_NOT_FOUND));
+        System.out.println(user);
+
+        return incProblemRecommendAlgorithm.recommend(user);
 
     }
 }
